+++ examples/browser/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @examples/browser
 
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @spotify-effect/browser@0.6.3
+  - @spotify-effect/core@0.6.3
+  - @spotify-effect/otel-node@0.6.3
+
 ## 0.0.7
 
 ### Patch Changes