+++ examples/solid/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @examples/solid
 
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @spotify-effect/browser@0.6.3
+  - @spotify-effect/core@0.6.3
+
 ## 0.0.6
 
 ### Patch Changes