+++ packages/otel-node/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @spotify-effect/otel-node
 
+## 0.6.3
+
 ## 0.6.2
 
 ### Patch Changes