+++ packages/spotify-effect/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @spotify-effect/core
 
+## 0.6.3
+
 ## 0.6.2
 
 ### Patch Changes